Move DisplayAbstractConcept into V2
